results/run-20230924-033218_LLMatic-zero_shot_donnemartin-system-design-oop-whatsapp-donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-dropbox-donnemartin-system-design-oop-twitter_0_5-runs_log/twitter/0/LLMatic/social.py
from models import Message

# Mock database
messages_db = {}


def send_message(sender_id, receiver_id, text):
	message = Message(len(messages_db) + 1, sender_id, receiver_id, text)
	messages_db[message.id] = message
	return message


def get_conversation(user1_id, user2_id):
	return sorted([message for message in messages_db.values() if (message.sender_id == user1_id and message.receiver_id == user2_id) or (message.sender_id == user2_id and message.receiver_id == user1_id)], key=lambda message: message.timestamp)

